backend/starlink-location/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
prometheus-client>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
numpy>=1.24.0
